order

+++ src/biz/IOrderBiz.java
@@ -7,5 +7,5 @@ import entity.Rorder;
 public interface IOrderBiz {
 	public List<Rorder> findOrderInfo(Integer uid, String condition, String type);
 	public int addOrder(String rorderid,Integer id,Integer uid,Integer state,Double total );
-	public int updateOrderState(Integer rorderid, Integer state);
+	public int updateOrderState(Double rorderid, Integer state);
 }
